+++ CHANGELOG.md
@@ -10,10 +10,12 @@ All notable changes to TazUO will be recorded here.
 ### Legion
 - Add sound API endpoints to LegionScripts - fpw
 - Added `API.ScriptName` and `API.ScriptPath`
+- Updated PSL browser UI and backend
 
 ### Assistant
 - Added skills tab to Legion Assistant - Coryigon
 - Organizer tab now shows graphic when hovering over the graphic art
+- Add Mobile outline option - Highlighting mobiles by notoriety
 
 ### Other
 - Move automatic py doc gen to tool usage